+++ lib/wrappers/linenoise/linenoise.nim
@@ -7,6 +7,8 @@
 #    distribution, for details about the copyright.
 #
 
+import std/private/rdstdin_impl
+
 type
   Completions* = object
     len*: csize_t
@@ -32,7 +34,7 @@ proc printKeyCodes*() {.importc: "linenoisePrintKeyCodes".}
 
 proc free*(s: cstring) {.importc: "free", header: "<stdlib.h>".}
 
-when defined(nimExperimentalLinenoiseExtra) and not defined(windows):
+when not defined(windows):
   # C interface
   type LinenoiseStatus = enum
     linenoiseStatus_ctrl_unknown
@@ -40,33 +42,6 @@ when defined(nimExperimentalLinenoiseExtra) and not defined(windows):
     linenoiseStatus_ctrl_D
 
   type LinenoiseData* = object
-    status: LinenoiseStatus
+    status*: LinenoiseStatus
 
-  proc linenoiseExtra(prompt: cstring, data: ptr LinenoiseData): cstring {.importc.}
-
-  # stable nim interface
-  type Status* = enum
-    lnCtrlUnkown
-    lnCtrlC
-    lnCtrlD
-
-  type ReadLineResult* = object
-    line*: string
-    status*: Status
-
-  proc readLineStatus*(prompt: string, result: var ReadLineResult) =
-    ## line editing API that allows returning the line entered and an indicator
-    ## of which control key was entered, allowing user to distinguish between
-    ## for example ctrl-C vs ctrl-D.
-    runnableExamples("-d:nimExperimentalLinenoiseExtra -r:off"):
-      var ret: ReadLineResult
-      while true:
-        readLineStatus("name: ", ret) # ctrl-D will exit, ctrl-C will go to next prompt
-        if ret.line.len > 0: echo ret.line
-        if ret.status == lnCtrlD: break
-      echo "exiting"
-    var data: LinenoiseData
-    let buf = linenoiseExtra(prompt, data.addr)
-    result.line = $buf
-    free(buf)
-    result.status = data.status.ord.Status
+  proc linenoiseExtra*(prompt: cstring, data: ptr LinenoiseData): cstring {.importc.}