+++ changelog.md
@@ -346,6 +346,10 @@
 
 - Added `dom.setInterval`, `dom.clearInterval` overloads.
 
+- `rdstdin.readLineFromStdin` (used by `nim secret`) now supports `^D` to indicate
+  end of file for platforms that support `linenoise`.
+
+
 ## Language changes
 
 - `nimscript` now handles `except Exception as e`.
@@ -459,7 +463,6 @@
   enforces that every symbol is written as it was declared, not enforcing
   the official Nim style guide. To be enabled, this has to be combined either
   with `--styleCheck:error` or `--styleCheck:hint`.
-
 
 
 ## Tool changes